AdditionalInfo
created a new widget MapDialog which allows for the popup of the map. Currently you can popup the map for the user to enter his location. Later on I will be adding the features to pass Placemarkers as a way for the provider to see the location entered for his accepted services.
I am still having problems extracting the location entered from the MapDialog since it is difficult to get data from a child Widget.